Refactor shadow runner tests
